+++ src/commonMain/kotlin/io/smallibs/pilin/control/extension/Comprehension.kt
@@ -1,7 +1,7 @@
 package io.smallibs.pilin.control.extension
 
 import io.smallibs.pilin.control.Monad
-import io.smallibs.pilin.delimited.Reflection
+import io.smallibs.pilin.control.extension.delimited.Reflection
 import io.smallibs.pilin.type.App
 
 class Comprehension<F, A>(private val monad: Monad.Core<F>) : Monad.Core<F> by monad {
@@ -12,7 +12,7 @@ class Comprehension<F, A>(private val monad: Monad.Core<F>) : Monad.Core<F> by m
         reflection.reflect(this)
 
     companion object {
-        suspend operator fun <F, A> invoke(monad: Monad.Core<F>, f: suspend Comprehension<F, A>.() -> A): App<F, A> =
+        suspend fun <F, A> run(monad: Monad.Core<F>, f: suspend Comprehension<F, A>.() -> A): App<F, A> =
             Comprehension<F, A>(monad).let { comprehension ->
                 comprehension.reflection.reify { comprehension.f() }
             }
