+++ src/commonMain/kotlin/io/smallibs/pilin/control/extension/delimited/thermometer/Frame.kt
@@ -1,4 +1,4 @@
-package io.smallibs.pilin.delimited.thermometer
+package io.smallibs.pilin.control.extension.delimited.thermometer
 
 sealed interface Frame {
     object Enter : Frame
